В этой ревизии внесены следующие изменения: 1) Теперь при импорте файлов вместо культурных настроек среды по умолчанию используются инвариантные 2) Теперь список кодов оплат берется из справочников, а не из предыдущих параметров расчета 3) Теперь после окочательного утверждения параметров расчета происходит обновление справочника кодов оплат 4) Теперь алгоритм приведения матриц учета окончательно переведен на System.Decimal 5) Теперь в справочник кодов оплат добавлено поле "Тип", иллюстрирующее отношение кода оплаты к резерву или командировке 6) Теперь в справочнике коды за номером "104" и "147" назначены командировочными по умолчанию, при этом все остальные назначены резервными 7) Исправлено множество мелких неполадок и недоработок, оптимизированы некоторые циклы, логики и т.п.

+++ src/NpoMash.Erm/NpoMash.Erm.Hrm/NpoMash.Erm.Hrm/Salary/HrmSalaryTaskMatrixReduction.cs
@@ -81,10 +81,10 @@ namespace NpoMash.Erm.Hrm.Salary {
         public class DepartmentItem : XPCustomObject {
             public Department Department;
             public DepartmentGroupDep Group;
-            public Int64 DepartmentPlan;
-            public Int64 MinimizeNumberOfDeviationsAlloc;
-            public Int64 MinimizeMaximumDeviationsAlloc;
-            public Int64 ProportionsMethodAlloc;
+            public Decimal DepartmentPlan;
+            public Decimal MinimizeNumberOfDeviationsAlloc;
+            public Decimal MinimizeMaximumDeviationsAlloc;
+            public Decimal ProportionsMethodAlloc;
             public IList<OrderItem> OrderItems = new List<OrderItem>();
             public DepartmentItem(Session session) : base(session) { }
         }
@@ -93,10 +93,10 @@ namespace NpoMash.Erm.Hrm.Salary {
         public class OrderItem : XPCustomObject {
             public fmCOrder Order;
             public FmCOrderTypeControl TypeControl;
-            public Int64 OrderPlan;
-            public Int64 MinimizeNumberOfDeviationsAlloc;
-            public Int64 MinimizeMaximumDeviationsAlloc;
-            public Int64 ProportionsMethodAlloc;
+            public Decimal OrderPlan;
+            public Decimal MinimizeNumberOfDeviationsAlloc;
+            public Decimal MinimizeMaximumDeviationsAlloc;
+            public Decimal ProportionsMethodAlloc;
             public IList<DepartmentItem> DepartmentItems = new List<DepartmentItem>();
             public OrderItem(Session session) : base(session) { }
         }
